DEVTOOLS-5011: Added tests
Testing has been added to the code and coverage is ~63%. Further
coverage is tricky to add, as mocking functions in go is a lot more work
than in a language like python, but plans are in place to do so.

Additionally, all code has been vetted by `go vet` and formatted by the
gofmt tool.