DEVTOOLS-5160: Fix URL display
The method by which the URL is displayed has been changed. The old
method of using a Log message does not display (this method is for
logging to the apt log in /var/). Attempting to use a Status message has
the URL cut off by apt, which makes it unsuitable for our purposes.
Fortunately, apt does not block using STDERR, so we can write it there.
The URLWriter is kept so we can prefix the message with a carriage
return such that we can overwrite the yellow [Working 0%] message apt
leaves.